requirements.txt
numpy~=1.23.5
pandas~=2.2.2
statsmodels~=0.14.2
scipy~=1.14.0
torch~=2.4.1
timm~=1.0.8
transformers~=4.44.0
sacred~=0.8.2
sentencepiece~=0.2.0
fairscale~=0.4.1
torchvision~=0.19.1  #+cu117
opencv-python~=4.10.0.84
scikit-learn~=1.5.1
omegaconf~=2.3.0
pytorch-lightning~=2.4.0
hydra-core~=1.3.2
wandb~=0.18.2
torchmetrics~=1.4.2